% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_nrads}
\alias{gut_nrads}
\title{Result of normalization of gut_otu_table with \code{max_rank = 400} and \code{average_over = 2000}.}
\format{
Contain the result of \code{\link{RADnormalization_matrix}(input = gut_otu_table,max_rank = 400,average_over = 2000)}
}
\usage{
data(gut_nrads)
}
\description{
Normalized rads created from \code{\link{gut_otu_table}} with \code{max_rank = 400} and \code{average_over = 2000}.
\code{\link{gut_otu_table}} has 18 gut samples from 3 individuals, prior, under and after using Ciprofloxacin (Cp)
antibiotic. Each row is the abundance of gut microbiome for one sample. The row names are as follows:
}
\details{
\itemize{
  \item A1, B1 and C1 are samples from individuals A, B and C 60 days prior using Cp.
  \item A2a is the sample from individual A 6 days prior using Cp.
  \item A2b is the sample from individual A 2 days prior using Cp.
  \item A2c, B2 and C2 are samples from individuals A, B and C one day prior using Cp.
  \item A3a is the sample from individual A 3 days after the first day of Cp administration.
  \item A3b, B3 and C3 are the samples from individual A, B and C 5 days after the first day of Cp administration.
  \item A4, B4 and C4 are the samples from individual A, B and C 33 days after the first day of Cp administration.
  \item A5, B5 and C5 are the samples from individual A, B and C 180 days after the first day of Cp administration.
}

points 1 and 2 are considered pre-Cp, points 3 are considered under-Cp and point 4 and 5 are considered post-Cp.

order of rows are similar to gut_otu_table
}
\keyword{datasets}
